// src/api/project_alarm.ts
export type ALARM_TYPE = number;
export const ALARM_TYPE_ISSUE_DEPEND_HIT: ALARM_TYPE = 0;
export const ALARM_TYPE_ISSUE_DEPEND_ALERT: ALARM_TYPE = 1;
export const ALARM_TYPE_ISSUE_DELAY_HIT: ALARM_TYPE = 2;
export const ALARM_TYPE_ISSUE_DELAY_ALERT: ALARM_TYPE = 3;
export const ALARM_TYPE_ISSUE_REOPEN_HIT: ALARM_TYPE = 4;
export const ALARM_TYPE_ISSUE_REOPEN_ALERT: ALARM_TYPE = 5;
// ...
export type IssueDependHitInfo = {
    issue_id: string;
    issue_type: ISSUE_TYPE;
};

export type IssueDependAlertInfo = {
    issue_id: string;
    issue_type: ISSUE_TYPE;
};

export type IssueDelayHitInfo = {
    issue_id: string;
    issue_type: ISSUE_TYPE;
};

export type IssueDelayAlertInfo = {
    issue_id: string;
    issue_type: ISSUE_TYPE;
};

export type IssueReOpenHitInfo = {
    issue_id: string;
    issue_type: ISSUE_TYPE;
};

export type IssueReOpenAlertInfo = {
    issue_id: string;
    issue_type: ISSUE_TYPE;
};

export type Content = {
    IssueDependHitInfo?: IssueDependHitInfo;
    IssueDependAlertInfo?: IssueDependAlertInfo;
    IssueDelayHitInfo?: IssueDelayHitInfo;
    IssueDelayAlertInfo?: IssueDelayAlertInfo;
    IssueReOpenHitInfo?: IssueReOpenHitInfo;
    IssueReOpenAlertInfo?: IssueReOpenAlertInfo;
};

export type Alarm = {
    alarm_id: string;
    project_id: string;
    time_stamp: number;
    alarm_type: ALARM_TYPE;
    title: string;
    content: Content;
};